feat(Cordova Hooks): Improve `Cordova hooks scripts` to work with Android SDK latest version
