+++ src/android/nearit.gradle
@@ -5,9 +5,17 @@ dependencies {
 }
 
 buildscript {
+	repositories {
+            jcenter()
+			mavenLocal()
+    }
+    
     dependencies {
+        classpath 'com.android.tools.build:gradle:+'
         classpath 'com.google.gms:google-services:3.0.0'
     }
 }
 
-apply plugin: 'com.google.gms.google-services'
+// apply plugin: 'com.google.gms.google-services'
+// class must be used instead of id(string) to be able to apply plugin from non-root gradle file
+apply plugin: com.google.gms.googleservices.GoogleServicesPlugin
